+++ src/mquery/mquery.js
@@ -60,7 +60,7 @@ class MQuery {
         if ('object' === type && !Array.isArray(arguments[0]))
             return this.merge(arguments[0]);
 
-        throw RxError.newRxTypeError('path must be a string or object', {
+        throw RxError.newRxTypeError('MQ1', {
             path: arguments[0]
         });
     }
@@ -225,7 +225,7 @@ class MQuery {
      */
     elemMatch() {
         if (null === arguments[0])
-            throw RxError.newRxTypeError('Invalid argument');
+            throw RxError.newRxTypeError('MQ2');
 
         let fn;
         let path;
@@ -246,7 +246,7 @@ class MQuery {
             path = arguments[0];
             criteria = arguments[1];
         } else
-            throw RxError.newRxTypeError('Invalid argument');
+            throw RxError.newRxTypeError('MQ2');
 
         if (fn) {
             criteria = new MQuery;
@@ -305,7 +305,7 @@ class MQuery {
             return this;
         }
 
-        throw RxError.newRxTypeError('Invalid sort() argument. Must be a string, object, or array.', {
+        throw RxError.newRxTypeError('MQ3', {
             args: arguments
         });
     }
@@ -323,7 +323,7 @@ class MQuery {
             return this;
 
         if (!MQuery.canMerge(source)) {
-            throw RxError.newRxTypeError('Invalid argument. Expected instanceof mquery or plain object', {
+            throw RxError.newRxTypeError('MQ4', {
                 source
             });
         }
@@ -392,11 +392,9 @@ class MQuery {
      */
     _ensurePath(method) {
         if (!this._path) {
-            throw RxError.newRxError(
-                'method must be used after where() when called with these arguments', {
-                    method
-                }
-            );
+            throw RxError.newRxError('MQ5', {
+                method
+            });
         }
     }
 }
@@ -433,7 +431,7 @@ class MQuery {
  */
 function push(opts, field, value) {
     if (Array.isArray(opts.sort)) {
-        throw RxError.newRxTypeError('Can\'t mix sort syntaxes. Use either array or object | .sort([[\'field\', 1], [\'test\', -1]]) | .sort({ field: 1, test: -1 })', {
+        throw RxError.newRxTypeError('MQ6', {
             opts,
             field,
             value
@@ -451,7 +449,7 @@ function push(opts, field, value) {
     const val = String(value || 1).toLowerCase();
     if (!/^(?:ascending|asc|descending|desc|1|-1)$/.test(val)) {
         if (Array.isArray(value)) value = '[' + value + ']';
-        throw RxError.newRxTypeError('Invalid sort value', {
+        throw RxError.newRxTypeError('MQ7', {
             field,
             value
         });
@@ -469,7 +467,7 @@ function push(opts, field, value) {
 function _pushArr(opts, field, value) {
     opts.sort = opts.sort || [];
     if (!Array.isArray(opts.sort)) {
-        throw RxError.newRxTypeError('Can\'t mix sort syntaxes. Use either array or object', {
+        throw RxError.newRxTypeError('MQ8', {
             opts,
             field,
             value