+++ src/plugins/encryption.js
@@ -44,18 +44,16 @@ export const prototypes = {
 };
 export const overwritable = {
     validatePassword: function(password) {
-        if (password && typeof password !== 'string'){
-            throw RxError.newRxTypeError('password is no string', {
+        if (password && typeof password !== 'string') {
+            throw RxError.newRxTypeError('EN1', {
                 password
             });
         }
         if (password && password.length < minPassLength) {
-            throw RxError.newRxError(
-                'validatePassword: min-length of password not complied', {
-                    minPassLength,
-                    password
-                }
-            );
+            throw RxError.newRxError('EN2', {
+                minPassLength,
+                password
+            });
         }
     }
 };