+++ src/typings/rx-error.d.ts
@@ -3,7 +3,17 @@ import { RxJsonSchema } from './rx-schema';
 export declare class RxError extends Error {
     readonly rxdb: boolean; // always true, use this to detect if its an rxdb-error
     readonly parameters: RxErrorParameters; // an object with parameters to use the programatically
+    readonly code: string; // error-code
+    readonly typeError: false; // true if is TypeError
 }
+
+export declare class RxTypeError extends TypeError {
+    readonly rxdb: boolean; // always true, use this to detect if its an rxdb-error
+    readonly parameters: RxErrorParameters; // an object with parameters to use the programatically
+    readonly code: string; // error-code
+    readonly typeError: true; // true if is TypeError
+}
+
 
 /**
  * this lists all possible parameters