+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Other:
   - Do not throw errors if the same plugin is added multiple times
   - Allow getting the collection via `RxDatabase().collection(name: string)`
   - Allow recreating the collection with different schema, if it has no documents
+  - Split out error-messages into separate [own plugin](https://pubkey.github.io/rxdb/custom-build.html#error-messages)
 
 ## 7.1.1 (November 27, 2017)